Refactor: move repo bundle path logic into Duffle internal paths module
