+++ src/duffle/duffle.paths.ts
@@ -4,6 +4,10 @@ export function credentialSetPath(name: string): string {
     return path.join(credentialSetDir(), name + '.yaml');
 }
 
+export function repoBundlePath(bundleRef: string): string {
+    return path.join(repositoriesDir(), localPath(bundleRef) + '.json');
+}
+
 function home(): string {
     const envHome = process.env['DUFFLE_HOME'];
     if (envHome) {
@@ -24,4 +28,15 @@ function osHome(): string {
 
 function credentialSetDir(): string {
     return path.join(home(), 'credentials');
+}
+
+function repositoriesDir(): string {
+    return path.join(home(), 'repositories');
+}
+
+function localPath(bundleRef: string): string {
+    const bits = bundleRef.split('/');
+    const last = bits.pop()!;
+    bits.push('bundles', last);
+    return bits.join('/');
 }
